FInal Changes
There is a bug where you turn on the first constellation, turn on another, turn off the first one, and then turn on the other, and the scene be will at square 1 but the second constellations song will still be playing. Other than that we good! Particle amounts and asteroid amounts could be changed but I think its cool